Move `FeedStoreSpecs` to its own file. `FeedStoreSpecs`  must be implemented by any `FeedStore` test implementation to guarantee it meets spec
